Stub for Kubernetes runtime
