+++ src/infractl/api/runtime.py
@@ -15,4 +15,10 @@ def runtime(*args, kind: str = 'prefect', **kwargs) -> infractl.base.Runtime:
 
         return infractl.base.Runtime(*args, kind=kind, **kwargs)
 
+    if kind == 'kubernetes':
+        # pylint: disable=import-outside-toplevel, unused-import, redefined-outer-name
+        import infractl.plugins.kubernetes_runtime.runtime  # noqa
+
+        return infractl.base.Runtime(*args, kind=kind, **kwargs)
+
     raise NotImplementedError(f'Runtime {kind} is not implemented')
