+++ src/infractl/plugins/kubernetes_runtime/runtime.py
@@ -1,0 +1,12 @@
+import infractl.base
+
+
+class KubernetesRuntimeImplementation(
+    infractl.base.RuntimeImplementation, registration_name='kubernetes'
+):
+    """Kubernetes runtime implementation."""
+
+    async def deploy(
+            self, program: infractl.base.program.Program, **kwargs
+    ) -> infractl.base.DeployedProgram:
+        raise NotImplementedError()
